cleanup, restructure
