+++ ots_eval/representatives_selection/example2.py
@@ -1,7 +1,7 @@
 from ots_eval.clustering.cots import COTS
 from representatives import Representatives
-from visualizations.plotly.plotter_3d import Plotter3d
-#from visualizations.seaborn2.plotter import Plotter
+#from visualizations.plotly.plotter_3d import Plotter
+from visualizations.seaborn import Plotter
 import pandas as pd
 
 
@@ -25,8 +25,7 @@ def get_representatives(df):
     return representatives
 
 
-def plot_result(df, representatives):
-    #pl=Plotter3d(df)
+def plot_result(df, representatives):   
     pl = Plotter(df)
     pl.add_representatives(representatives)
     fig=pl.generate_fig()
