+++ visualizations/seaborn/__init__.py
@@ -1,0 +1,1 @@
+from visualizations.seaborn.Plotter import Plotter
